movie api cleanup

+++ express/modules/movies/movie.controller.js
@@ -1,14 +1,7 @@
 // CRUD operations
 const movieModel = require("./movie.model");
 const { slugger } = require("../../utils/text");
-
-// movie create {create}
-// movie list {list}
-// get one movie {getById}
-// update releaseDate {updateReleaseDate}
-// update movieDetails {update}
-// delete movie {remove} // check whether movie's ticket has been sold
-// update seats {updateSeats}
+const moment = require("moment");
 
 const create = async (payload) => {
   // create slug from title eg : The Invincibles slug->the-invincibles {use slugify package in /utils}
@@ -30,7 +23,9 @@ const getBySlug = async (slug) => {
 };
 
 const update = async (slug, payload) => {
-  // TODO check if movie exists and also use slug instead of id
+  if (payload.title) {
+    payload.slug = slugger(payload?.title);
+  }
   return await movieModel.findOneAndUpdate({ slug }, payload, { new: true });
 };
 
@@ -40,13 +35,28 @@ const updateReleaseDate = async (slug, payload) => {
 };
 
 const updateSeats = async (slug, payload) => {
+  const movie = await movieModel.findOne({ slug });
+  if (!movie) throw new Error("Movie doesnt exist");
   // check if the movie seats are less than defined number(process.env.MIN_SEATS)
+  if (payload.seats < process.env.MIN_SEATS)
+    throw new Error(`Movie seats cant be less than ${process.env.MIN_SEATS}`);
   return await movieModel.findOneAndUpdate({ slug }, payload, { new: true });
 };
 
 const remove = async (slug) => {
-  // movie's ticket must not have been sold
+  const movie = await movieModel.findOne({ slug });
+  if (!movie) throw new Error("Movie doesnt exist");
+  // TODO movie's ticket must not have been sold
+
   // the movie should not be ongoing (should not be in between relase and end dates)
+  if (
+    moment(movie?.releaseDate).isBefore(moment()) &&
+    moment(movie?.endDate).isAfter(moment())
+  ) {
+    throw new Error(
+      "Cannot delete movie since movie is currently in releasing phase"
+    );
+  }
   return await movieModel.deleteOne({ slug });
 };
